Refactor @truffle/preserve-fs
- Extract common types/interfaces into separate types.ts file
- Slightly reorder fs.ts to accommodate reading the file top-down
- Slight changes due to refactoring of @truffle/preserve
- Split up test file into fs.fixture.ts and fs.test.ts
