+++ packages/preserve-fs/test/fs.fixture.ts
@@ -1,23 +1,18 @@
-import fs from "fs-extra";
-import path from "path";
-import tmp from "tmp-promise";
-
 import * as Preserve from "@truffle/preserve";
-import { Loader } from "..";
 
-interface File {
+export interface File {
   path: string;
   content: string;
 }
 
-interface Test {
+export interface Test {
   name: string;
   files: File[];
   targeted: string; // path to target (will be prefixed)
-  expected: Preserve.Targets.Thunked.Target;
+  expected: Preserve.Targets.Stringified.Target;
 }
 
-const tests: Test[] = [
+export const tests: Test[] = [
   {
     name: "single-file",
     files: [
@@ -109,53 +104,3 @@ const tests: Test[] = [
     }
   }
 ];
-
-const writeFile = async (fullPath: string, content: string): Promise<void> => {
-  // ensure directory exists for file
-  await fs.ensureDir(path.dirname(fullPath));
-
-  await fs.promises.writeFile(fullPath, content);
-};
-
-describe("Loader", () => {
-  let workspace: tmp.DirectoryResult;
-
-  beforeAll(async () => {
-    workspace = await tmp.dir();
-  });
-
-  afterAll(async () => {
-    await workspace.cleanup();
-  });
-
-  for (const { name, files, targeted, expected } of tests) {
-    describe(`test: ${name}`, () => {
-      beforeAll(async () => {
-        for (const file of files) {
-          const fullPath = path.join(workspace.path, name, file.path);
-
-          await writeFile(fullPath, file.content);
-        }
-      });
-
-      it("returns correct target", async () => {
-        const fullPath = path.join(workspace.path, name, targeted);
-
-        const loader = new Loader();
-
-        const target: Preserve.Target = await Preserve.Controllers.run(
-          {
-            method: loader.load.bind(loader)
-          },
-          {
-            path: fullPath
-          }
-        );
-
-        const thunked = await Preserve.Targets.thunk(target);
-
-        expect(thunked).toEqual(expected);
-      });
-    });
-  }
-});
